Add files via upload
I think this works. I don't really know how to test it. When I look in the Serial Monitor it says that the note goes off every 1 second +/- a couple milliseconds. I think the milliseconds are from the time it takes to send Serial data from the arduino to my computer.